Finally moving common date_arg to icenet.cli, along with setup_logging
