+++ icenet/cli.py
@@ -1,0 +1,43 @@
+import datetime as dt
+from functools import wraps
+import logging
+import re
+
+
+def setup_logging(func,
+                  log_format="[%(asctime)-17s :%(levelname)-8s] - %(message)s"):
+
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        parsed_args = func(*args, **kwargs)
+        level = logging.INFO
+
+        if hasattr(parsed_args, "verbose") and parsed_args.verbose:
+            level = logging.DEBUG
+
+        logging.basicConfig(
+            level=level,
+            format=log_format,
+            datefmt="%d-%m-%y %T",
+        )
+
+        # TODO: better way of handling these on a case by case basis
+        logging.getLogger("cdsapi").setLevel(logging.WARNING)
+        logging.getLogger("matplotlib").setLevel(logging.WARNING)
+        logging.getLogger("matplotlib.pyplot").setLevel(logging.WARNING)
+        logging.getLogger("requests").setLevel(logging.WARNING)
+        logging.getLogger("tensorflow").setLevel(logging.WARNING)
+        logging.getLogger("urllib3").setLevel(logging.WARNING)
+        return parsed_args
+
+    return wrapper
+
+
+def date_arg(string: str) -> object:
+    """
+
+    :param string:
+    :return:
+    """
+    date_match = re.search(r"(\d{4})-(\d{1,2})-(\d{1,2})", string)
+    return dt.date(*[int(s) for s in date_match.groups()])
