+++ icenet/data/dataset.py
@@ -10,7 +10,7 @@ import pandas as pd
 from icenet.data.datasets.utils import SplittingMixin
 from icenet.data.loader import IceNetDataLoaderFactory
 from icenet.data.producers import DataCollection
-from icenet.utils import setup_logging
+from icenet.cli import setup_logging
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.WARNING)
