+++ icenet/process/local.py
@@ -5,8 +5,8 @@ import shutil
 
 import xarray as xr
 
-from icenet.process.utils import date_arg, destination_filename
-from icenet.utils import setup_logging
+from icenet.cli import date_arg, setup_logging
+from icenet.process.utils import destination_filename
 
 
 @setup_logging
